+++ packages/functions/test/utils.ts
@@ -16,9 +16,15 @@
  */
 
 import { FirebaseOptions, FirebaseApp } from '@firebase/app-types';
-import { Provider, ComponentContainer } from '@firebase/component';
+import {
+  Provider,
+  ComponentContainer,
+  Component,
+  ComponentType
+} from '@firebase/component';
 import { FirebaseAuthInternalName } from '@firebase/auth-interop-types';
 import { FirebaseMessagingName } from '@firebase/messaging-types';
+import { AppCheckInternalComponentName } from '@firebase/app-check-interop-types';
 import { Service } from '../src/api/service';
 import nodeFetch from 'node-fetch';
 
@@ -41,6 +47,12 @@ export function makeFakeApp(options: FirebaseOptions = {}): FirebaseApp {
   };
 }
 
+export function getFetchImpl(): typeof fetch {
+  return typeof window !== 'undefined'
+    ? fetch.bind(window)
+    : (nodeFetch as any);
+}
+
 export function createTestService(
   app: FirebaseApp,
   regionOrCustomDomain?: string,
@@ -51,14 +63,30 @@ export function createTestService(
   messagingProvider = new Provider<FirebaseMessagingName>(
     'messaging',
     new ComponentContainer('test')
-  )
+  ),
+  fakeAppCheckToken = 'dummytoken',
+  fetchImpl = getFetchImpl()
 ): Service {
-  const fetchImpl: typeof fetch =
-    typeof window !== 'undefined' ? fetch.bind(window) : (nodeFetch as any);
+  const appCheckProvider = new Provider<AppCheckInternalComponentName>(
+    'app-check-internal',
+    new ComponentContainer('test')
+  );
+  appCheckProvider.setComponent(
+    new Component(
+      'app-check-internal',
+      () => {
+        return {
+          getToken: () => Promise.resolve({ token: fakeAppCheckToken })
+        } as any; // eslint-disable-line @typescript-eslint/no-explicit-any
+      },
+      ComponentType.PRIVATE
+    )
+  );
   const functions = new Service(
     app,
     authProvider,
     messagingProvider,
+    appCheckProvider,
     regionOrCustomDomain,
     fetchImpl
   );