+++ packages/storage/src/service.ts
@@ -23,6 +23,7 @@ import { XhrIoPool } from './implementation/xhriopool';
 import { Reference, _getChild } from './reference';
 import { Provider } from '@firebase/component';
 import { FirebaseAuthInternalName } from '@firebase/auth-interop-types';
+import { AppCheckInternalComponentName } from '@firebase/app-check-interop-types';
 import {
   FirebaseApp,
   FirebaseOptions,
@@ -166,6 +167,13 @@ export class StorageService implements _FirebaseService {
      */
     readonly app: FirebaseApp,
     readonly _authProvider: Provider<FirebaseAuthInternalName>,
+    /**
+     * @internal
+     */
+    readonly _appCheckProvider: Provider<AppCheckInternalComponentName>,
+    /**
+     * @internal
+     */
     readonly _pool: XhrIoPool,
     readonly _url?: string,
     readonly _firebaseVersion?: string
@@ -244,6 +252,19 @@ export class StorageService implements _FirebaseService {
     return null;
   }
 
+  async _getAppCheckToken(): Promise<string | null> {
+    const appCheck = this._appCheckProvider.getImmediate({ optional: true });
+    if (appCheck) {
+      const result = await appCheck.getToken();
+      // TODO: What do we want to do if there is an error getting the token?
+      // Context: appCheck.getToken() will never throw even if an error happened. In the error case, a dummy token will be
+      // returned along with an error field describing the error. In general, we shouldn't care about the error condition and just use
+      // the token (actual or dummy) to send requests.
+      return result.token;
+    }
+    return null;
+  }
+
   /**
    * Stop running requests and prevent more from being created.
    */
@@ -268,13 +289,15 @@ export class StorageService implements _FirebaseService {
    */
   _makeRequest<T>(
     requestInfo: RequestInfo<T>,
-    authToken: string | null
+    authToken: string | null,
+    appCheckToken: string | null
   ): Request<T> {
     if (!this._deleted) {
       const request = makeRequest(
         requestInfo,
         this._appId,
         authToken,
+        appCheckToken,
         this._pool,
         this._firebaseVersion
       );
@@ -288,5 +311,16 @@ export class StorageService implements _FirebaseService {
     } else {
       return new FailRequest(appDeleted());
     }
+  }
+
+  async makeRequestWithTokens<T>(
+    requestInfo: RequestInfo<T>
+  ): Promise<Request<T>> {
+    const [authToken, appCheckToken] = await Promise.all([
+      this._getAuthToken(),
+      this._getAppCheckToken()
+    ]);
+
+    return this._makeRequest(requestInfo, authToken, appCheckToken);
   }
 }