+++ packages/storage/test/unit/service.compat.test.ts
@@ -25,6 +25,7 @@ import { StorageService } from '../../src/service';
 import { FirebaseApp } from '@firebase/app-types';
 import { Provider } from '@firebase/component';
 import { FirebaseAuthInternalName } from '@firebase/auth-interop-types';
+import { AppCheckInternalComponentName } from '@firebase/app-check-interop-types';
 import { TestingXhrIo } from './xhrio';
 import { Headers } from '../../src/implementation/xhrio';
 
@@ -40,12 +41,13 @@ function makeGsUrl(child: string = ''): string {
 function makeService(
   app: FirebaseApp,
   authProvider: Provider<FirebaseAuthInternalName>,
+  appCheckProvider: Provider<AppCheckInternalComponentName>,
   pool: XhrIoPool,
   url?: string
 ): StorageServiceCompat {
   const storageServiceCompat: StorageServiceCompat = new StorageServiceCompat(
     app,
-    new StorageService(app, authProvider, pool, url)
+    new StorageService(app, authProvider, appCheckProvider, pool, url)
   );
   return storageServiceCompat;
 }
@@ -55,6 +57,7 @@ describe('Firebase Storage > Service', () => {
     const service = makeService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
+      testShared.fakeAppCheckTokenProvider,
       xhrIoPool
     );
     it('Root refs point to the right place', () => {
@@ -89,6 +92,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool,
         'gs://foo-bar.appspot.com'
       );
@@ -99,6 +103,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool,
         `http://${DEFAULT_HOST}/v1/b/foo-bar.appspot.com/o`
       );
@@ -109,6 +114,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool,
         `https://${DEFAULT_HOST}/v1/b/foo-bar.appspot.com/o`
       );
@@ -120,6 +126,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool,
         'foo-bar.appspot.com'
       );
@@ -130,6 +137,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool,
         'foo-bar.appspot.com'
       );
@@ -141,6 +149,7 @@ describe('Firebase Storage > Service', () => {
         makeService(
           testShared.fakeApp,
           testShared.fakeAuthProvider,
+          testShared.fakeAppCheckTokenProvider,
           xhrIoPool,
           'gs://bucket/object/'
         );
@@ -153,6 +162,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         fakeAppGs,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool
       );
       expect(service.ref().toString()).to.equal('gs://mybucket/');
@@ -161,13 +171,19 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         fakeAppGsEndingSlash,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         xhrIoPool
       );
       expect(service.ref().toString()).to.equal('gs://mybucket/');
     });
     it('Throws when config bucket is gs:// with an object path', () => {
       testShared.assertThrows(() => {
-        makeService(fakeAppInvalidGs, testShared.fakeAuthProvider, xhrIoPool);
+        makeService(
+          fakeAppInvalidGs,
+          testShared.fakeAuthProvider,
+          testShared.fakeAppCheckTokenProvider,
+          xhrIoPool
+        );
       }, 'storage/invalid-default-bucket');
     });
   });
@@ -189,6 +205,7 @@ describe('Firebase Storage > Service', () => {
       const service = makeService(
         testShared.fakeApp,
         testShared.fakeAuthProvider,
+        testShared.fakeAppCheckTokenProvider,
         testShared.makePool(newSend)
       );
       service.useEmulator('test.host.org', 1234);
@@ -200,6 +217,7 @@ describe('Firebase Storage > Service', () => {
     const service = makeService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
+      testShared.fakeAppCheckTokenProvider,
       xhrIoPool
     );
     it('Works with gs:// URLs', () => {
@@ -258,6 +276,7 @@ GOOG4-RSA-SHA256`
     const service = makeService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
+      testShared.fakeAppCheckTokenProvider,
       xhrIoPool
     );
     describe('ref', () => {
@@ -307,6 +326,7 @@ GOOG4-RSA-SHA256`
     const service = makeService(
       testShared.fakeApp,
       testShared.fakeAuthProvider,
+      testShared.fakeAppCheckTokenProvider,
       xhrIoPool
     );
     it('In-flight requests are canceled when the service is deleted', async () => {